+++ components/button.tsx
@@ -1,27 +1,27 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import clsx from "clsx";
-import { ClipLoader } from "react-spinners";
-
 
 type Variant = "solid" | "outline";
 type ColorSchema = "blue" | "red" | "green";
-type Size = "sm" | "md" | "full" | "Large";
 
-interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+interface ButtonProps {
   title: string;
   isDisabled?: boolean;
   variant?: Variant;
   colorSchema: ColorSchema;
-  width?: Size;
-  height?: Size;
+  width?: string;
+  height?: string;
   isLoading?: boolean;
 }
 
-const Button: React.FC<ButtonProps> = ({
+const Button: React.FC<
+  ButtonProps & React.ButtonHTMLAttributes<HTMLButtonElement>
+> = ({
   title,
   isDisabled = false,
   variant = "solid",
   colorSchema,
-  width = "md",
+  width = "full",
   height = "md",
   isLoading = false,
   ...props
@@ -30,42 +30,34 @@ const Button: React.FC<ButtonProps> = ({
     <button
       {...props}
       disabled={isDisabled}
-      className={clsx(
-        "rounded-full border capitalize flex items-center justify-center",
-        {
-          // Warna solid
-          "bg-blue-500 text-white": colorSchema === "blue" && variant === "solid",
-          "bg-red-500 text-white": colorSchema === "red" && variant === "solid",
-          "bg-green-500 text-white": colorSchema === "green" && variant === "solid",
-
-          // Warna outline
-          "border-blue-500 text-blue-500 bg-transparent":
-            colorSchema === "blue" && variant === "outline",
-          "border-red-500 text-red-500 bg-transparent":
-            colorSchema === "red" && variant === "outline",
-          "border-green-500 text-green-500 bg-transparent":
-            colorSchema === "green" && variant === "outline",
-
-          // Disabled
-          "opacity-50 cursor-not-allowed": isDisabled,
-
-          // Width
-          "w-20": width === "sm",
-          "w-24": width === "md",
-          "w-32" : width === "Large",
-          "w-full": width === "full",
-
-          // Height
-          "h-8": height === "sm",
-          "h-10": height === "md",
-          "h-12": height === "full",
-        }
-      )}
+      className={clsx(` rounded border w-20  capitalize`, {
+        "bg-blue-500 text-white": colorSchema === "blue" && variant === "solid",
+        "border-blue-500 text-blue-500":
+          colorSchema === "blue" && variant === "outline",
+        "bg-red-500 text-white": colorSchema === "red" && variant === "solid",
+        "border-red-500 text-red-500 ":
+          colorSchema === "red" && variant === "outline",
+        "bg-green-500 text-white":
+          colorSchema === "green" && variant === "solid",
+        "border-green-500 text-green-500":
+          colorSchema === "green" && variant === "outline",
+        "opacity-25": isDisabled,
+        "w-40": width === "xl",
+        "w-full": width === "full",
+        "h-8": width === "sm",
+        "w-20": width === "md",
+      })}
     >
-      {isLoading ? (
-        <ClipLoader color="#fff" size={20} />
-      ) : (
+      {!isLoading ? (
         title
+      ) : (
+        <p>Loading ...</p>
+        // <ClipLoader
+        //   color={"#36d7b7"}
+        //   size={20}
+        //   aria-label="Loading Spinner"
+        //   data-testid="loader"
+        // />
       )}
     </button>
   );
